// Frontend/src/app/processos/models/informacao-ultima-operacao.ts
export class InformacoesUltimaOperacao {
    nomeUsuarioLog: string;
    codUsuarioLog: string;
    nroProcessoCartorioD: string;
    codEstadoD: string;
    codVaraD: number;
    nomTipoV: string;
    nomTipoVaraD: string;
    operacao: string;
    nomeUsuari: string;
    codUsuario: string;
    nomeComarcaD: string;
    nomParteOrgaoD: string;
    dataLog: string;
}